backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.11.0",
    "beautifulsoup4>=4.14.3",
    "dj-database-url>=3.0.1",
    "django==5.2.9",
    "django-cors-headers>=4.9.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-ai-slim[groq]>=1.27.0",
    "pyjwt==2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlparse==0.5.4",
]
